+++ egui_glium/CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the `egui_glium` integration will be noted in this file.
 ## Unreleased
 * [Fix minimize on Windows](https://github.com/emilk/egui/issues/518)
 * Change `drag_and_drop_support` to `false` by default (Windows only). See <https://github.com/emilk/egui/issues/598>.
+* Don't restore window position on Windows, because the position would sometimes be invalid.
 
 
 ## 0.13.1 - 2021-06-24